Use loggers and fix extract_slices.py
