+++ new_data_processing/extract_slices.py
@@ -1,181 +1,17 @@
-import os
 import json
-import clang.cindex
-import clang.enumerations
-import csv
-import numpy as np
-import re 
-import warnings
-from tqdm import tqdm
-import sys, argparse
+import logging
+from pathlib import Path
+
+import tqdm
 from graphviz import Digraph
 
-warnings.filterwarnings('ignore')
-# set the config
-try:
-    clang.cindex.Config.set_library_path("/work/LAS/weile-lab/benjis/weile-lab/thesis/ReVeal/clang+llvm-6.0.1-x86_64-linux-gnu-ubuntu-16.04/lib")
-    clang.cindex.Config.set_library_file('/work/LAS/weile-lab/benjis/weile-lab/thesis/ReVeal/clang+llvm-6.0.1-x86_64-linux-gnu-ubuntu-16.04/lib/libclang.so.6.0')
-except:
-    pass
+logger = logging.getLogger(__name__)
 
-l_funcs = ['StrNCat', 'getaddrinfo', '_ui64toa', 'fclose', 'pthread_mutex_lock', 'gets_s', 'sleep', 
-           '_ui64tot', 'freopen_s', '_ui64tow', 'send', 'lstrcat', 'HMAC_Update', '__fxstat', 'StrCatBuff', 
-           '_mbscat', '_mbstok_s', '_cprintf_s', 'ldap_search_init_page', 'memmove_s', 'ctime_s', 'vswprintf', 
-           'vswprintf_s', '_snwprintf', '_gmtime_s', '_tccpy', '*RC6*', '_mbslwr_s', 'random', 
-           '__wcstof_internal', '_wcslwr_s', '_ctime32_s', 'wcsncat*', 'MD5_Init', '_ultoa', 
-           'snprintf', 'memset', 'syslog', '_vsnprintf_s', 'HeapAlloc', 'pthread_mutex_destroy', 
-           'ChangeWindowMessageFilter', '_ultot', 'crypt_r', '_strupr_s_l', 'LoadLibraryExA', '_strerror_s', 
-           'LoadLibraryExW', 'wvsprintf', 'MoveFileEx', '_strdate_s', 'SHA1', 'sprintfW', 'StrCatNW', 
-           '_scanf_s_l', 'pthread_attr_init', '_wtmpnam_s', 'snscanf', '_sprintf_s_l', 'dlopen', 
-           'sprintfA', 'timed_mutex', 'OemToCharA', 'ldap_delete_ext', 'sethostid', 'popen', 'OemToCharW', 
-           '_gettws', 'vfork', '_wcsnset_s_l', 'sendmsg', '_mbsncat', 'wvnsprintfA', 'HeapFree', '_wcserror_s', 
-           'realloc', '_snprintf*', 'wcstok', '_strncat*', 'StrNCpy', '_wasctime_s', 'push*', '_lfind_s', 
-           'CC_SHA512', 'ldap_compare_ext_s', 'wcscat_s', 'strdup', '_chsize_s', 'sprintf_s', 'CC_MD4_Init', 
-           'wcsncpy', '_wfreopen_s', '_wcsupr_s', '_searchenv_s', 'ldap_modify_ext_s', '_wsplitpath', 
-           'CC_SHA384_Final', 'MD2', 'RtlCopyMemory', 'lstrcatW', 'MD4', 'MD5', '_wcstok_s_l', '_vsnwprintf_s', 
-           'ldap_modify_s', 'strerror', '_lsearch_s', '_mbsnbcat_s', '_wsplitpath_s', 'MD4_Update', '_mbccpy_s', 
-           '_strncpy_s_l', '_snprintf_s', 'CC_SHA512_Init', 'fwscanf_s', '_snwprintf_s', 'CC_SHA1', 'swprintf', 
-           'fprintf', 'EVP_DigestInit_ex', 'strlen', 'SHA1_Init', 'strncat', '_getws_s', 'CC_MD4_Final', 
-           'wnsprintfW', 'lcong48', 'lrand48', 'write', 'HMAC_Init', '_wfopen_s', 'wmemchr', '_tmakepath', 
-           'wnsprintfA', 'lstrcpynW', 'scanf_s', '_mbsncpy_s_l', '_localtime64_s', 'fstream.open', '_wmakepath', 
-           'Connection.open', '_tccat', 'valloc', 'setgroups', 'unlink', 'fstream.put', 'wsprintfA', '*SHA1*', 
-           '_wsearchenv_s', 'ualstrcpyA', 'CC_MD5_Update', 'strerror_s', 'HeapCreate', 'ualstrcpyW', '__xstat', 
-           '_wmktemp_s', 'StrCatChainW', 'ldap_search_st', '_mbstowcs_s_l', 'ldap_modify_ext', '_mbsset_s', 
-           'strncpy_s', 'move', 'execle', 'StrCat', 'xrealloc', 'wcsncpy_s', '_tcsncpy*', 'execlp', 
-           'RIPEMD160_Final', 'ldap_search_s', 'EnterCriticalSection', '_wctomb_s_l', 'fwrite', '_gmtime64_s', 
-           'sscanf_s', 'wcscat', '_strupr_s', 'wcrtomb_s', 'VirtualLock', 'ldap_add_ext_s', '_mbscpy', 
-           '_localtime32_s', 'lstrcpy', '_wcsncpy*', 'CC_SHA1_Init', '_getts', '_wfopen', '__xstat64', 
-           'strcoll', '_fwscanf_s_l', '_mbslwr_s_l', 'RegOpenKey', 'makepath', 'seed48', 'CC_SHA256', 
-           'sendto', 'execv', 'CalculateDigest', 'memchr', '_mbscpy_s', '_strtime_s', 'ldap_search_ext_s', 
-           '_chmod', 'flock', '__fxstat64', '_vsntprintf', 'CC_SHA256_Init', '_itoa_s', '__wcserror_s', 
-           '_gcvt_s', 'fstream.write', 'sprintf', 'recursive_mutex', 'strrchr', 'gethostbyaddr', '_wcsupr_s_l', 
-           'strcspn', 'MD5_Final', 'asprintf', '_wcstombs_s_l', '_tcstok', 'free', 'MD2_Final', 'asctime_s', 
-           '_alloca', '_wputenv_s', '_wcsset_s', '_wcslwr_s_l', 'SHA1_Update', 'filebuf.sputc', 'filebuf.sputn', 
-           'SQLConnect', 'ldap_compare', 'mbstowcs_s', 'HMAC_Final', 'pthread_condattr_init', '_ultow_s', 'rand', 
-           'ofstream.put', 'CC_SHA224_Final', 'lstrcpynA', 'bcopy', 'system', 'CreateFile*', 'wcscpy_s', 
-           '_mbsnbcpy*', 'open', '_vsnwprintf', 'strncpy', 'getopt_long', 'CC_SHA512_Final', '_vsprintf_s_l', 
-           'scanf', 'mkdir', '_localtime_s', '_snprintf', '_mbccpy_s_l', 'memcmp', 'final', '_ultoa_s', 
-           'lstrcpyW', 'LoadModule', '_swprintf_s_l', 'MD5_Update', '_mbsnset_s_l', '_wstrtime_s', '_strnset_s', 
-           'lstrcpyA', '_mbsnbcpy_s', 'mlock', 'IsBadHugeWritePtr', 'copy', '_mbsnbcpy_s_l', 'wnsprintf', 
-           'wcscpy', 'ShellExecute', 'CC_MD4', '_ultow', '_vsnwprintf_s_l', 'lstrcpyn', 'CC_SHA1_Final', 
-           'vsnprintf', '_mbsnbset_s', '_i64tow', 'SHA256_Init', 'wvnsprintf', 'RegCreateKey', 'strtok_s', 
-           '_wctime32_s', '_i64toa', 'CC_MD5_Final', 'wmemcpy', 'WinExec', 'CreateDirectory*', 
-           'CC_SHA256_Update', '_vsnprintf_s_l', 'jrand48', 'wsprintf', 'ldap_rename_ext_s', 'filebuf.open', 
-           '_wsystem', 'SHA256_Update', '_cwscanf_s', 'wsprintfW', '_sntscanf', '_splitpath', 'fscanf_s', 
-           'strpbrk', 'wcstombs_s', 'wscanf', '_mbsnbcat_s_l', 'strcpynA', 'pthread_cond_init', 'wcsrtombs_s', 
-           '_wsopen_s', 'CharToOemBuffA', 'RIPEMD160_Update', '_tscanf', 'HMAC', 'StrCCpy', 'Connection.connect', 
-           'lstrcatn', '_mbstok', '_mbsncpy', 'CC_SHA384_Update', 'create_directories', 'pthread_mutex_unlock', 
-           'CFile.Open', 'connect', '_vswprintf_s_l', '_snscanf_s_l', 'fputc', '_wscanf_s', '_snprintf_s_l', 
-           'strtok', '_strtok_s_l', 'lstrcatA', 'snwscanf', 'pthread_mutex_init', 'fputs', 'CC_SHA384_Init', 
-           '_putenv_s', 'CharToOemBuffW', 'pthread_mutex_trylock', '__wcstoul_internal', '_memccpy', 
-           '_snwprintf_s_l', '_strncpy*', 'wmemset', 'MD4_Init', '*RC4*', 'strcpyW', '_ecvt_s', 'memcpy_s', 
-           'erand48', 'IsBadHugeReadPtr', 'strcpyA', 'HeapReAlloc', 'memcpy', 'ldap_rename_ext', 'fopen_s', 
-           'srandom', '_cgetws_s', '_makepath', 'SHA256_Final', 'remove', '_mbsupr_s', 'pthread_mutexattr_init', 
-           '__wcstold_internal', 'StrCpy', 'ldap_delete', 'wmemmove_s', '_mkdir', 'strcat', '_cscanf_s_l', 
-           'StrCAdd', 'swprintf_s', '_strnset_s_l', 'close', 'ldap_delete_ext_s', 'ldap_modrdn', 'strchr', 
-           '_gmtime32_s', '_ftcscat', 'lstrcatnA', '_tcsncat', 'OemToChar', 'mutex', 'CharToOem', 'strcpy_s', 
-           'lstrcatnW', '_wscanf_s_l', '__lxstat64', 'memalign', 'MD2_Init', 'StrCatBuffW', 'StrCpyN', 'CC_MD5', 
-           'StrCpyA', 'StrCatBuffA', 'StrCpyW', 'tmpnam_r', '_vsnprintf', 'strcatA', 'StrCpyNW', '_mbsnbset_s_l', 
-           'EVP_DigestInit', '_stscanf', 'CC_MD2', '_tcscat', 'StrCpyNA', 'xmalloc', '_tcslen', '*MD4*', 
-           'vasprintf', 'strxfrm', 'chmod', 'ldap_add_ext', 'alloca', '_snscanf_s', 'IsBadWritePtr', 'swscanf_s', 
-           'wmemcpy_s', '_itoa', '_ui64toa_s', 'EVP_DigestUpdate', '__wcstol_internal', '_itow', 'StrNCatW', 
-           'strncat_s', 'ualstrcpy', 'execvp', '_mbccat', 'EVP_MD_CTX_init', 'assert', 'ofstream.write', 
-           'ldap_add', '_sscanf_s_l', 'drand48', 'CharToOemW', 'swscanf', '_itow_s', 'RIPEMD160_Init', 
-           'CopyMemory', 'initstate', 'getpwuid', 'vsprintf', '_fcvt_s', 'CharToOemA', 'setuid', 'malloc', 
-           'StrCatNA', 'strcat_s', 'srand', 'getwd', '_controlfp_s', 'olestrcpy', '__wcstod_internal', 
-           '_mbsnbcat', 'lstrncat', 'des_*', 'CC_SHA224_Init', 'set*', 'vsprintf_s', 'SHA1_Final', '_umask_s', 
-           'gets', 'setstate', 'wvsprintfW', 'LoadLibraryEx', 'ofstream.open', 'calloc', '_mbstrlen', 
-           '_cgets_s', '_sopen_s', 'IsBadStringPtr', 'wcsncat_s', 'add*', 'nrand48', 'create_directory', 
-           'ldap_search_ext', '_i64toa_s', '_ltoa_s', '_cwscanf_s_l', 'wmemcmp', '__lxstat', 'lstrlen', 
-           'pthread_condattr_destroy', '_ftcscpy', 'wcstok_s', '__xmknod', 'pthread_attr_destroy', 'sethostname', 
-           '_fscanf_s_l', 'StrCatN', 'RegEnumKey', '_tcsncpy', 'strcatW', 'AfxLoadLibrary', 'setenv', 'tmpnam', 
-           '_mbsncat_s_l', '_wstrdate_s', '_wctime64_s', '_i64tow_s', 'CC_MD4_Update', 'ldap_add_s', '_umask', 
-           'CC_SHA1_Update', '_wcsset_s_l', '_mbsupr_s_l', 'strstr', '_tsplitpath', 'memmove', '_tcscpy', 
-           'vsnprintf_s', 'strcmp', 'wvnsprintfW', 'tmpfile', 'ldap_modify', '_mbsncat*', 'mrand48', 'sizeof', 
-           'StrCatA', '_ltow_s', '*desencrypt*', 'StrCatW', '_mbccpy', 'CC_MD2_Init', 'RIPEMD160', 'ldap_search', 
-           'CC_SHA224', 'mbsrtowcs_s', 'update', 'ldap_delete_s', 'getnameinfo', '*RC5*', '_wcsncat_s_l', 
-           'DriverManager.getConnection', 'socket', '_cscanf_s', 'ldap_modrdn_s', '_wopen', 'CC_SHA256_Final', 
-           '_snwprintf*', 'MD2_Update', 'strcpy', '_strncat_s_l', 'CC_MD5_Init', 'mbscpy', 'wmemmove', 
-           'LoadLibraryW', '_mbslen', '*alloc', '_mbsncat_s', 'LoadLibraryA', 'fopen', 'StrLen', 'delete', 
-           '_splitpath_s', 'CreateFileTransacted*', 'MD4_Final', '_open', 'CC_SHA384', 'wcslen', 'wcsncat', 
-           '_mktemp_s', 'pthread_mutexattr_destroy', '_snwscanf_s', '_strset_s', '_wcsncpy_s_l', 'CC_MD2_Final', 
-           '_mbstok_s_l', 'wctomb_s', 'MySQL_Driver.connect', '_snwscanf_s_l', '*_des_*', 'LoadLibrary', 
-           '_swscanf_s_l', 'ldap_compare_s', 'ldap_compare_ext', '_strlwr_s', 'GetEnvironmentVariable', 
-           'cuserid', '_mbscat_s', 'strspn', '_mbsncpy_s', 'ldap_modrdn2', 'LeaveCriticalSection', 'CopyFile', 
-           'getpwd', 'sscanf', 'creat', 'RegSetValue', 'ldap_modrdn2_s', 'CFile.Close', '*SHA_1*', 
-           'pthread_cond_destroy', 'CC_SHA512_Update', '*RC2*', 'StrNCatA', '_mbsnbcpy', '_mbsnset_s', 
-           'crypt', 'excel', '_vstprintf', 'xstrdup', 'wvsprintfA', 'getopt', 'mkstemp', '_wcsnset_s', 
-           '_stprintf', '_sntprintf', 'tmpfile_s', 'OpenDocumentFile', '_mbsset_s_l', '_strset_s_l', 
-           '_strlwr_s_l', 'ifstream.open', 'xcalloc', 'StrNCpyA', '_wctime_s', 'CC_SHA224_Update', '_ctime64_s', 
-           'MoveFile', 'chown', 'StrNCpyW', 'IsBadReadPtr', '_ui64tow_s', 'IsBadCodePtr', 'getc', 
-           'OracleCommand.ExecuteOracleScalar', 'AccessDataSource.Insert', 'IDbDataAdapter.FillSchema', 
-           'IDbDataAdapter.Update', 'GetWindowText*', 'SendMessage', 'SqlCommand.ExecuteNonQuery', 'streambuf.sgetc', 
-           'streambuf.sgetn', 'OracleCommand.ExecuteScalar', 'SqlDataSource.Update', '_Read_s', 'IDataAdapter.Fill', 
-           '_wgetenv', '_RecordsetPtr.Open*', 'AccessDataSource.Delete', 'Recordset.Open*', 'filebuf.sbumpc', 'DDX_*', 
-           'RegGetValue', 'fstream.read*', 'SqlCeCommand.ExecuteResultSet', 'SqlCommand.ExecuteXmlReader', 'main', 
-           'streambuf.sputbackc', 'read', 'm_lpCmdLine', 'CRichEditCtrl.Get*', 'istream.putback', 
-           'SqlCeCommand.ExecuteXmlReader', 'SqlCeCommand.BeginExecuteXmlReader', 'filebuf.sgetn', 
-           'OdbcDataAdapter.Update', 'filebuf.sgetc', 'SQLPutData', 'recvfrom', 'OleDbDataAdapter.FillSchema', 
-           'IDataAdapter.FillSchema', 'CRichEditCtrl.GetLine', 'DbDataAdapter.Update', 'SqlCommand.ExecuteReader', 
-           'istream.get', 'ReceiveFrom', '_main', 'fgetc', 'DbDataAdapter.FillSchema', 'kbhit', 'UpdateCommand.Execute*', 
-           'Statement.execute', 'fgets', 'SelectCommand.Execute*', 'getch', 'OdbcCommand.ExecuteNonQuery', 
-           'CDaoQueryDef.Execute', 'fstream.getline', 'ifstream.getline', 'SqlDataAdapter.FillSchema', 
-           'OleDbCommand.ExecuteReader', 'Statement.execute*', 'SqlCeCommand.BeginExecuteNonQuery', 
-           'OdbcCommand.ExecuteScalar', 'SqlCeDataAdapter.Update', 'sendmessage', 'mysqlpp.DBDriver', 'fstream.peek', 
-           'Receive', 'CDaoRecordset.Open', 'OdbcDataAdapter.FillSchema', '_wgetenv_s', 'OleDbDataAdapter.Update', 
-           'readsome', 'SqlCommand.BeginExecuteXmlReader', 'recv', 'ifstream.peek', '_Main', '_tmain', '_Readsome_s', 
-           'SqlCeCommand.ExecuteReader', 'OleDbCommand.ExecuteNonQuery', 'fstream.get', 'IDbCommand.ExecuteScalar', 
-           'filebuf.sputbackc', 'IDataAdapter.Update', 'streambuf.sbumpc', 'InsertCommand.Execute*', 'RegQueryValue', 
-           'IDbCommand.ExecuteReader', 'SqlPipe.ExecuteAndSend', 'Connection.Execute*', 'getdlgtext', 'ReceiveFromEx', 
-           'SqlDataAdapter.Update', 'RegQueryValueEx', 'SQLExecute', 'pread', 'SqlCommand.BeginExecuteReader', 'AfxWinMain', 
-           'getchar', 'istream.getline', 'SqlCeDataAdapter.Fill', 'OleDbDataReader.ExecuteReader', 'SqlDataSource.Insert', 
-           'istream.peek', 'SendMessageCallback', 'ifstream.read*', 'SqlDataSource.Select', 'SqlCommand.ExecuteScalar', 
-           'SqlDataAdapter.Fill', 'SqlCommand.BeginExecuteNonQuery', 'getche', 'SqlCeCommand.BeginExecuteReader', 'getenv', 
-           'streambuf.snextc', 'Command.Execute*', '_CommandPtr.Execute*', 'SendNotifyMessage', 'OdbcDataAdapter.Fill', 
-           'AccessDataSource.Update', 'fscanf', 'QSqlQuery.execBatch', 'DbDataAdapter.Fill', 'cin', 
-           'DeleteCommand.Execute*', 'QSqlQuery.exec', 'PostMessage', 'ifstream.get', 'filebuf.snextc', 
-           'IDbCommand.ExecuteNonQuery', 'Winmain', 'fread', 'getpass', 'GetDlgItemTextCCheckListBox.GetCheck', 
-           'DISP_PROPERTY_EX', 'pread64', 'Socket.Receive*', 'SACommand.Execute*', 'SQLExecDirect', 
-           'SqlCeDataAdapter.FillSchema', 'DISP_FUNCTION', 'OracleCommand.ExecuteNonQuery', 'CEdit.GetLine', 
-           'OdbcCommand.ExecuteReader', 'CEdit.Get*', 'AccessDataSource.Select', 'OracleCommand.ExecuteReader', 
-           'OCIStmtExecute', 'getenv_s', 'DB2Command.Execute*', 'OracleDataAdapter.FillSchema', 'OracleDataAdapter.Fill', 
-           'CComboBox.Get*', 'SqlCeCommand.ExecuteNonQuery', 'OracleCommand.ExecuteOracleNonQuery', 'mysqlpp.Query', 
-           'istream.read*', 'CListBox.GetText', 'SqlCeCommand.ExecuteScalar', 'ifstream.putback', 'readlink', 
-           'CHtmlEditCtrl.GetDHtmlDocument', 'PostThreadMessage', 'CListCtrl.GetItemText', 'OracleDataAdapter.Update', 
-           'OleDbCommand.ExecuteScalar', 'stdin', 'SqlDataSource.Delete', 'OleDbDataAdapter.Fill', 'fstream.putback', 
-           'IDbDataAdapter.Fill', '_wspawnl', 'fwprintf', 'sem_wait', '_unlink', 'ldap_search_ext_sW', 'signal', 'PQclear', 
-           'PQfinish', 'PQexec', 'PQresultStatus']
+from constants import l_funcs
 
-def read_csv(csv_file_path):
-    data = []
-    with open(csv_file_path) as fp:
-        header = fp.readline()
-        header = header.strip()
-        h_parts = [hp.strip() for hp in header.split('\t')]
-        for line in fp:
-            line = line.strip()
-            instance = {}
-            lparts = line.split('\t')
-            for i, hp in enumerate(h_parts):
-                if i < len(lparts):
-                    content = lparts[i].strip()
-                else:
-                    content = ''
-                instance[hp] = content
-            data.append(instance)
-        return data
+import clang_stuff
 
-
-def read_code_file(file_path):
-    code_lines = {}
-    with open(file_path) as fp:
-        for ln, line in enumerate(fp):
-            assert isinstance(line, str)
-            line = line.strip()
-            if '//' in line:
-                line = line[:line.index('//')]
-            code_lines[ln + 1] = line
-        return code_lines
+from utils import read_csv
 
 
 def extract_nodes_with_location_info(nodes):
@@ -202,13 +38,14 @@ def extract_nodes_with_location_info(nodes):
     return node_indices, node_ids, line_numbers, node_id_to_line_number
     pass
 
+
 def create_adjacency_list(line_numbers, node_id_to_line_numbers, edges, data_dependency_only=False):
     adjacency_list = {}
     for ln in set(line_numbers):
         adjacency_list[ln] = [set(), set()]
     for edge in edges:
         edge_type = edge['type'].strip()
-        if True :#edge_type in ['IS_AST_PARENT', 'FLOWS_TO']:
+        if True:  # edge_type in ['IS_AST_PARENT', 'FLOWS_TO']:
             start_node_id = edge['start'].strip()
             end_node_id = edge['end'].strip()
             if start_node_id not in node_id_to_line_numbers.keys() or end_node_id not in node_id_to_line_numbers.keys():
@@ -216,9 +53,9 @@ def create_adjacency_list(line_numbers, node_id_to_line_numbers, edges, data_dep
             start_ln = node_id_to_line_numbers[start_node_id]
             end_ln = node_id_to_line_numbers[end_node_id]
             if not data_dependency_only:
-                if edge_type == 'CONTROLS': #Control Flow edges
+                if edge_type == 'CONTROLS':  # Control Flow edges
                     adjacency_list[start_ln][0].add(end_ln)
-            if edge_type == 'REACHES': # Data Flow edges
+            if edge_type == 'REACHES':  # Data Flow edges
                 adjacency_list[start_ln][1].add(end_ln)
     return adjacency_list
 
@@ -277,89 +114,7 @@ def create_backward_slice(adjacency_list, line_no):
     inverted_adjacency_list = invert_graph(adjacency_list)
     return create_forward_slice(inverted_adjacency_list, line_no)
 
-class Tokenizer:
-    # creates the object, does the inital parse
-    def __init__(self, path, tokenizer_type='original'):
-        self.index = clang.cindex.Index.create()
-        self.tu = self.index.parse(path)
-        self.path = self.extract_path(path)
-        self.symbol_table = {}
-        self.symbol_count = 1
-        self.tokenizer_type = tokenizer_type
 
-    # To output for split_functions, must have same path up to last two folders
-    def extract_path(self, path):
-        return "".join(path.split("/")[:-2])
-
-    
-    def full_tokenize_cursor(self, cursor):
-        tokens = cursor.get_tokens()
-        result = []
-        for token in tokens:
-            if token.kind.name == "COMMENT":
-                continue
-            if token.kind.name == "LITERAL":
-                result += self.process_literal(token)
-                continue
-            if token.kind.name == "IDENTIFIER":
-                result += ["ID"]
-                continue
-            result += [token.spelling]
-        return result
-
-    def full_tokenize(self):
-        cursor = self.tu.cursor
-        return self.full_tokenize_cursor(cursor)
-
-    def process_literal(self, literal):
-        cursor_kind = clang.cindex.CursorKind
-        kind = literal.cursor.kind
-        if kind == cursor_kind.INTEGER_LITERAL:
-            return literal.spelling
-        if kind == cursor_kind.FLOATING_LITERAL:
-            return literal.spelling
-        if kind == cursor_kind.IMAGINARY_LITERAL:
-            return ["NUM"]       
-        if kind == cursor_kind.STRING_LITERAL:
-            return ["STRING"]
-        sp = literal.spelling
-        if re.match('[0-9]+', sp) is not None:
-            return sp
-        return ["LITERAL"]
-
-    def split_functions(self, method_only):
-        results = []
-        cursor_kind = clang.cindex.CursorKind
-        cursor = self.tu.cursor
-        for c in cursor.get_children():
-            filename = c.location.file.name if c.location.file != None else "NONE"
-            extracted_path = self.extract_path(filename)
-
-            if (c.kind == cursor_kind.CXX_METHOD or (method_only == False and c.kind == cursor_kind.FUNCTION_DECL)) and extracted_path == self.path:
-                name = c.spelling
-                tokens = self.full_tokenize_cursor(c)
-                filename = filename.split("/")[-1]
-                results += [tokens]
-
-        return results
-    
-
-def tokenize(file_text):
-    try:
-        c_file = open('/tmp/test1.c', 'w')
-        c_file.write(file_text)
-        c_file.close()
-        tok = Tokenizer('/tmp/test1.c')
-        results = tok.split_functions(False)
-        return ' '.join(results[0])
-    except:
-        return None
-
-def read_file(path):
-    with open(path) as f:
-        lines = f.readlines()
-        return ' '.join(lines)
-    
 def extract_line_number(idx, nodes):
     while idx >= 0:
         c_node = nodes[idx]
@@ -374,171 +129,179 @@ def extract_line_number(idx, nodes):
         idx -= 1
     return -1
 
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--project', help='name of project for differentiating files', default='chrome_debian')
-    parser.add_argument('--input', help='directory where raw code and parsed are stored', default='../data/chrome_debian')
-    parser.add_argument('--text_in', help='path to cfg_full_text_files', default='../data/ggnn_input/')
-    parser.add_argument('--output', help='output path for sliced full data json', default='../data/')
-    args = parser.parse_args()
 
-    split_dir = args.input + 'raw_code/'
-    parsed = args.input + 'parsed/'
+def get_slices(combined_graph, nodes):
+    call_lines = set()
+    array_lines = set()
+    ptr_lines = set()
+    arithmatic_lines = set()
 
-    ggnn_json_data = json.load(open(args.text_in + args.project + '_cfg_full_text_files.json'))
-    files = [d['file_name'] for d in ggnn_json_data]
-    print(f'Number of Input Files: {len(files)}')
-
-    all_data = []
-        
-    for i, file_name  in enumerate(files):
-        label = file_name.strip()[:-2].split('_')[-1 if '_refactored' not in file_name else -2]
-        code_text = read_file(split_dir + file_name.strip())
-        
-        nodes_file_path = parsed + file_name.strip() + '/nodes.csv'
-        edges_file_path = parsed + file_name.strip() + '/edges.csv'
-        nc = open(nodes_file_path)
-        nodes_file = csv.DictReader(nc, delimiter='\t')
-        nodes = [node for node in nodes_file]
-        call_lines = set()
-        array_lines = set()
-        ptr_lines = set()
-        arithmatic_lines = set()
-        
-        if len(nodes) == 0:
-            continue
-        
-        for node_idx, node in enumerate(nodes):
-            ntype = node['type'].strip()
-            if ntype == 'CallExpression':
-                function_name = nodes[node_idx + 1]['code']
-                if function_name  is None or function_name.strip() == '':
-                    continue
-                if function_name.strip() in l_funcs:
-                    line_no = extract_line_number(node_idx, nodes)
-                    if line_no > 0:
-                        call_lines.add(line_no)
-            elif ntype == 'ArrayIndexing':
+    for node_idx, node in enumerate(nodes):
+        ntype = node['type'].strip()
+        if ntype == 'CallExpression':
+            function_name = nodes[node_idx + 1]['code']
+            if function_name is None or function_name.strip() == '':
+                continue
+            if function_name.strip() in l_funcs:
                 line_no = extract_line_number(node_idx, nodes)
                 if line_no > 0:
-                    array_lines.add(line_no)
-            elif ntype == 'PtrMemberAccess':
-                line_no = extract_line_number(node_idx, nodes)
-                if line_no > 0:
-                    ptr_lines.add(line_no)
-            elif node['operator'].strip() in ['+', '-', '*', '/']:
-                line_no = extract_line_number(node_idx, nodes)
-                if line_no > 0:
-                    arithmatic_lines.add(line_no)
-            
+                    call_lines.add(line_no)
+        elif ntype == 'ArrayIndexing':
+            line_no = extract_line_number(node_idx, nodes)
+            if line_no > 0:
+                array_lines.add(line_no)
+        elif ntype == 'PtrMemberAccess':
+            line_no = extract_line_number(node_idx, nodes)
+            if line_no > 0:
+                ptr_lines.add(line_no)
+        elif node['operator'].strip() in ['+', '-', '*', '/']:
+            line_no = extract_line_number(node_idx, nodes)
+            if line_no > 0:
+                arithmatic_lines.add(line_no)
+    array_slices = []
+    array_slices_bdir = []
+    call_slices = []
+    call_slices_bdir = []
+    arith_slices = []
+    arith_slices_bdir = []
+    ptr_slices = []
+    ptr_slices_bdir = []
+    all_slices = []
+    all_keys = set()
+    _keys = set()
+    for slice_ln in call_lines:
+        forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
+        backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
+        all_slice_lines = forward_sliced_lines
+        all_slice_lines.extend(backward_sliced_lines)
+        all_slice_lines = sorted(list(set(all_slice_lines)))
+        key = ' '.join([str(i) for i in all_slice_lines])
+        if key not in _keys:
+            call_slices.append(backward_sliced_lines)
+            call_slices_bdir.append(all_slice_lines)
+            _keys.add(key)
+        if key not in all_keys:
+            all_slices.append(all_slice_lines)
+            all_keys.add(key)
+    _keys = set()
+    for slice_ln in array_lines:
+        forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
+        backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
+        all_slice_lines = forward_sliced_lines
+        all_slice_lines.extend(backward_sliced_lines)
+        all_slice_lines = sorted(list(set(all_slice_lines)))
+        key = ' '.join([str(i) for i in all_slice_lines])
+        if key not in _keys:
+            array_slices.append(backward_sliced_lines)
+            array_slices_bdir.append(all_slice_lines)
+            _keys.add(key)
+        if key not in all_keys:
+            all_slices.append(all_slice_lines)
+            all_keys.add(key)
+    _keys = set()
+    for slice_ln in arithmatic_lines:
+        forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
+        backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
+        all_slice_lines = forward_sliced_lines
+        all_slice_lines.extend(backward_sliced_lines)
+        all_slice_lines = sorted(list(set(all_slice_lines)))
+        key = ' '.join([str(i) for i in all_slice_lines])
+        if key not in _keys:
+            arith_slices.append(backward_sliced_lines)
+            arith_slices_bdir.append(all_slice_lines)
+            _keys.add(key)
+        if key not in all_keys:
+            all_slices.append(all_slice_lines)
+            all_keys.add(key)
+    _keys = set()
+    for slice_ln in ptr_lines:
+        forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
+        backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
+        all_slice_lines = forward_sliced_lines
+        all_slice_lines.extend(backward_sliced_lines)
+        all_slice_lines = sorted(list(set(all_slice_lines)))
+        key = ' '.join([str(i) for i in all_slice_lines])
+        if key not in _keys:
+            ptr_slices.append(backward_sliced_lines)
+            ptr_slices_bdir.append(all_slice_lines)
+            _keys.add(key)
+        if key not in all_keys:
+            all_slices.append(all_slice_lines)
+            all_keys.add(key)
+    return {
+        'call_slices_vd': call_slices,
+        'call_slices_sy': call_slices_bdir,
+        'array_slices_vd': array_slices,
+        'array_slices_sy': array_slices_bdir,
+        'arith_slices_vd': arith_slices,
+        'arith_slices_sy': arith_slices_bdir,
+        'ptr_slices_vd': ptr_slices,
+        'ptr_slices_sy': ptr_slices_bdir,
+    }
+
+
+def preprocess(input_data, input_dir, output_dir, project, store=False, vuld_syse=False):
+    if isinstance(input_data, Path):
+        input_data = json.load(open(input_data))
+
+    code_dir = input_dir / 'raw_code'
+    parsed_dir = input_dir / 'parsed'
+    assert code_dir.exists(), code_dir
+    assert parsed_dir.exists(), parsed_dir
+    logger.info(f'Number of Input Files: {len(input_data)}')
+
+    if vuld_syse:
+        clang_stuff.setup()
+
+    output_data = []
+    for i, d in tqdm.tqdm(enumerate(input_data)):
+        file_name = d["file_name"]
+        label = d["label"]
+        code = d["code"]
+
+        # Sanity checks
+        assert file_name == file_name.strip()
+        assert label in (0, 1)
+
+        data_instance = {
+            'file_path': str(code_dir / file_name),
+            'code': code,
+            'label': int(label)
+        }
+
+        # Load Joern output
+        nodes_file_path = parsed_dir / file_name / 'nodes.csv'
+        edges_file_path = parsed_dir / file_name / 'edges.csv'
         nodes = read_csv(nodes_file_path)
         edges = read_csv(edges_file_path)
+        if len(nodes) == 0:
+            logger.debug(f'Skipping node {i} ({file_name}) because len(nodes) == 0')
+            continue
+
+        # Make graph data structure
         node_indices, node_ids, line_numbers, node_id_to_ln = extract_nodes_with_location_info(nodes)
         adjacency_list = create_adjacency_list(line_numbers, node_id_to_ln, edges, False)
         combined_graph = combine_control_and_data_adjacents(adjacency_list)
-        
-        array_slices = []
-        array_slices_bdir = []
-        call_slices = []
-        call_slices_bdir = []
-        arith_slices = []
-        arith_slices_bdir = []
-        ptr_slices = []
-        ptr_slices_bdir = []
-        all_slices = []
-        
-        
-        all_keys = set()
-        _keys = set()
-        for slice_ln in call_lines:
-            forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
-            backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
-            all_slice_lines = forward_sliced_lines
-            all_slice_lines.extend(backward_sliced_lines)
-            all_slice_lines = sorted(list(set(all_slice_lines)))
-            key = ' '.join([str(i) for i in all_slice_lines])
-            if key not in _keys:
-                call_slices.append(backward_sliced_lines)
-                call_slices_bdir.append(all_slice_lines)
-                _keys.add(key)
-            if key not in all_keys:
-                all_slices.append(all_slice_lines)
-                all_keys.add(key)
-                
-        _keys = set()
-        for slice_ln in array_lines:
-            forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
-            backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
-            all_slice_lines = forward_sliced_lines
-            all_slice_lines.extend(backward_sliced_lines)
-            all_slice_lines = sorted(list(set(all_slice_lines)))
-            key = ' '.join([str(i) for i in all_slice_lines])
-            if key not in _keys:
-                array_slices.append(backward_sliced_lines)
-                array_slices_bdir.append(all_slice_lines)
-                _keys.add(key)
-            if key not in all_keys:
-                all_slices.append(all_slice_lines)
-                all_keys.add(key)
-        
-        _keys = set()
-        for slice_ln in arithmatic_lines:
-            forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
-            backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
-            all_slice_lines = forward_sliced_lines
-            all_slice_lines.extend(backward_sliced_lines)
-            all_slice_lines = sorted(list(set(all_slice_lines)))
-            key = ' '.join([str(i) for i in all_slice_lines])
-            if key not in _keys:
-                arith_slices.append(backward_sliced_lines)
-                arith_slices_bdir.append(all_slice_lines)
-                _keys.add(key)
-            if key not in all_keys:
-                all_slices.append(all_slice_lines)
-                all_keys.add(key)
-        
-        _keys = set()
-        for slice_ln in ptr_lines:
-            forward_sliced_lines = create_forward_slice(combined_graph, slice_ln)
-            backward_sliced_lines = create_backward_slice(combined_graph, slice_ln)
-            all_slice_lines = forward_sliced_lines
-            all_slice_lines.extend(backward_sliced_lines)
-            all_slice_lines = sorted(list(set(all_slice_lines)))
-            key = ' '.join([str(i) for i in all_slice_lines])
-            if key not in _keys:
-                ptr_slices.append(backward_sliced_lines)
-                ptr_slices_bdir.append(all_slice_lines)
-                _keys.add(key)
-            if key not in all_keys:
-                all_slices.append(all_slice_lines)
-                all_keys.add(key)
-                
-        t_code = tokenize(code_text)
+
+        # Tokenize code
+        code_text = ' '.join(code.splitlines(keepends=True))
+        t_code = clang_stuff.tokenize(code_text)
         if t_code is None:
+            logger.debug(f'Skipping node {i} ({file_name}) because t_code is None')
             continue
-        data_instance = {
-            'file_path': split_dir + file_name.strip(),
-            'code' : code_text,
-            'tokenized': t_code,
-            'call_slices_vd': call_slices,
-            'call_slices_sy': call_slices_bdir,
-            'array_slices_vd': array_slices,
-            'array_slices_sy': array_slices_bdir,
-            'arith_slices_vd': arith_slices,
-            'arith_slices_sy': arith_slices_bdir,
-            'ptr_slices_vd': ptr_slices,
-            'ptr_slices_sy': ptr_slices_bdir,
-            'label': int(label)
-        }
-        all_data.append(data_instance)
-        
-        if i % 1000 == 0:
-            print(i, len(call_slices), len(call_slices_bdir), 
-                len(array_slices), len(array_slices_bdir), 
-                len(arith_slices), len(arith_slices_bdir), sep='\t')
+        data_instance["tokenized"] = t_code
 
+        # Get slices for VulDeePecker/SySeVR
+        if vuld_syse:
+            vuld_syse_slice_data = get_slices(combined_graph, nodes)
+            data_instance.update(vuld_syse_slice_data)
 
-    print(len(all_data))
-    output_file = open(args.output + args.project + '_full_data_with_slices.json', 'w')
-    json.dump(all_data, output_file)
-    output_file.close()
+        # Return data instance
+        output_data.append(data_instance)
+    logger.info(len(output_data), 'items')
+
+    if store:
+        with open(output_dir / (project + '_full_data_with_slices.json'), 'w') as output_file:
+            json.dump(output_data, output_file)
+        logger.info(f'saved output File to {output_file}')
+    return output_data
