+++ new_data_processing/utils.py
@@ -1,0 +1,18 @@
+def read_csv(csv_file_path):
+    data = []
+    with open(csv_file_path) as fp:
+        header = fp.readline()
+        header = header.strip()
+        h_parts = [hp.strip() for hp in header.split('\t')]
+        for line in fp:
+            line = line.strip()
+            instance = {}
+            lparts = line.split('\t')
+            for i, hp in enumerate(h_parts):
+                if i < len(lparts):
+                    content = lparts[i].strip()
+                else:
+                    content = ''
+                instance[hp] = content
+            data.append(instance)
+        return data
